Everything works. Just need GUI.
One weird, seemingly random occurrence happens when trying to add a player to the game... Sometimes the player is null? It happens a lot when the clientID is 23, but it doesn't happen every time and it also happens on other numbers as well. Resetting the server typically fixes it so we can really just not worry about it for the presentation.